+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Bump version of `brick/math`
 - Bump version of `phpstan/phpstan`
 - Ensure opaque paths always round trip per [whatwg/url#784](https://github.com/whatwg/url/issues/784) and [whatwg/url#844](https://github.com/whatwg/url/pull/844)
+- Make port state override return failure more for URLPattern per [whatwg/url#866](https://github.com/whatwg/url/pull/866) and [whatwg/urlpattern#266](https://github.com/whatwg/urlpattern/issues/266)
 
 ### Internals